+++ app/src/main/java/com/example/heaapp/model/food/Sugars.java
@@ -12,13 +12,13 @@ public class Sugars {
     private String unit;
     @SerializedName("per_hundred")
     @Expose
-    private Integer perHundred;
+    private Double perHundred;
     @SerializedName("per_portion")
     @Expose
     private Double perPortion;
     @SerializedName("per_day")
     @Expose
-    private Integer perDay;
+    private Double perDay;
 
     public NameTranslations____ getNameTranslations() {
         return nameTranslations;
@@ -36,11 +36,11 @@ public class Sugars {
         this.unit = unit;
     }
 
-    public Integer getPerHundred() {
+    public Double getPerHundred() {
         return perHundred;
     }
 
-    public void setPerHundred(Integer perHundred) {
+    public void setPerHundred(Double perHundred) {
         this.perHundred = perHundred;
     }
 
@@ -52,11 +52,11 @@ public class Sugars {
         this.perPortion = perPortion;
     }
 
-    public Integer getPerDay() {
+    public Double getPerDay() {
         return perDay;
     }
 
-    public void setPerDay(Integer perDay) {
+    public void setPerDay(Double perDay) {
         this.perDay = perDay;
     }
